requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.0.0
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
psycopg2==2.9.9
python-dotenv==1.0.1
PyJWT==2.8.0
requests==2.31.0
